+++ src/zakuro.cpp
@@ -358,17 +358,44 @@ int Zakuro::Run() {
         webrtc::EnableMedia(dependencies);
       };
 
-  if (!vc_config.openh264.empty()) {
-    context_config.video_codec_factory_config.capability_config.openh264_path = vc_config.openh264;
-    auto& preference =
-      context_config.video_codec_factory_config.preference.emplace();
-    auto capability = sora::GetVideoCodecCapability(
+  context_config.video_codec_factory_config.capability_config
+      .get_custom_engines = []() {
+    // NopVideoDecoder
+    sora::VideoCodecCapability::Engine engine(
+        sora::VideoCodecImplementation::kCustom_1);
+    engine.parameters.custom_engine_name = "NopVideoDecoder";
+    engine.codecs.emplace_back(webrtc::kVideoCodecVP8, false, true);
+    engine.codecs.emplace_back(webrtc::kVideoCodecVP9, false, true);
+    engine.codecs.emplace_back(webrtc::kVideoCodecH264, false, true);
+    engine.codecs.emplace_back(webrtc::kVideoCodecH265, false, false);
+    engine.codecs.emplace_back(webrtc::kVideoCodecAV1, false, false);
+    return std::vector<sora::VideoCodecCapability::Engine>{engine};
+  };
+
+  auto capability = sora::GetVideoCodecCapability(
       context_config.video_codec_factory_config.capability_config);
-    preference.Merge(sora::CreateVideoCodecPreferenceFromImplementation(
+  auto& preference =
+      context_config.video_codec_factory_config.preference.emplace();
+  preference.Merge(sora::CreateVideoCodecPreferenceFromImplementation(
       capability, sora::VideoCodecImplementation::kInternal));
+  if (!vc_config.openh264.empty()) {
+    context_config.video_codec_factory_config.capability_config.openh264_path =
+        vc_config.openh264;
     preference.Merge(sora::CreateVideoCodecPreferenceFromImplementation(
-      capability, sora::VideoCodecImplementation::kCiscoOpenH264));
+        capability, sora::VideoCodecImplementation::kCiscoOpenH264));
   }
+  preference.Merge(sora::CreateVideoCodecPreferenceFromImplementation(
+      capability, sora::VideoCodecImplementation::kCustom_1));
+  context_config.video_codec_factory_config.create_video_decoder =
+      [](sora::VideoCodecImplementation implementation,
+         const sora::VideoCodecCapabilityConfig& capability_config,
+         webrtc::VideoCodecType type) {
+        if (implementation == sora::VideoCodecImplementation::kCustom_1) {
+          return std::make_unique<NopVideoDecoder>();
+        } else {
+          throw "Invalid implementation";
+        }
+      };
 
   vc_config.context = sora::SoraClientContext::Create(context_config);